# spec/swagger.yaml
swagger: '2.0'
info:
  title: SuggestGrid
  version: 0.1.32-SNAPSHOT
  description: Personalization made Simple
  x-codegen-settings:
    appendContentHeaders: true
    generateAsyncCode: true
    useMethodPrefix: false
    useModelPostfix: false
    useEnumPostfix: false
    useConstructorsForConfig: false
    iOSUseAppInfoPlist: true
    androidUseAppManifest: true
    collectParameters: false
    csharpDefaultNamespace: SuggestGrid
    javaDefaultPackageName: com.suggestgrid
    brandLabel: SuggestGrid
    userAgent: SUGGESTGRID
    projectName: SuggestGrid
    enableAdditionalModelProperties: false
    preserveParameterOrder: true
    enablePHPComposerVersionString: true
schemes:
  - https
securityDefinitions:
  basicAuth:
    type: basic
    description: HTTP Basic Authentication.
security:
  - basicAuth: []
consumes:
  - application/json
produces:
  - application/json
...
definitions:
...
  Metadata:
    type: object
    description: |
      Metadata of a user or an item.
      Id parameter is required for all metadata.
      Metadata may have other parameters with string, number, or boolean values.
    properties:
      id:
        type: string
        description: |
          The id of the metadata of a user or an item.
    additionalProperties:
      type:
        - string
        - integer
        - number
        - boolean
    required:
      - id
  # Request Bodies
...
  MessageResponse:
    type: object
    description: Basic response with a message for success responses.
    properties:
      message:
        type: string
        description: Message of the response.
...
  ErrorResponse:
    type: object
    description: Error response.
    properties:
      error_text:
        type: string
        description: Message of the response.
      error_description:
        type: string
        description: Description of the response.
      error_uri:
        type: string
        description: The URI of the error for more details.
  DetailedErrorResponse:
    type: object
    description: Error response.
    properties:
      error_text:
        type: string
        description: Message of the response.
      error_description:
        type: string
        description: Description of the response.
      error_uri:
        type: string
        description: The URI of the error for more details.
      error_details:
        type: string
        description: Details of the response.
...
  BulkPostResponse:
    type: object
    description: |
      Bulk error response.
      Returned on all bulk post requests regardless of the number of erroneous or successful posts.
    properties:
      message:
        type: string
        description: Message of the response.
      errors:
        type: array
        items:
          $ref: '#/definitions/BulkPostError'
  BulkPostError:
    type: object
    description: Error response returned due to bulk post schema validations.
    properties:
      message:
        type: string
        description: Message of the response.
      value:
        type: object
        description: The cause of the error.
      error:
        type: object
        description: Programatic description of the error.
  UsersResponse:
    type: object
    description: Get users response.
    properties:
      count:
        type: integer
        format: int64
        description: The number of users in the response.
      total_count:
        type: integer
        format: int64
        description: The total number of users available.
      users:
        type: array
        items:
          $ref: '#/definitions/Metadata'
  ItemsResponse:
    type: object
    description: Get items response.
    properties:
      count:
        type: integer
        format: int64
        description: The number of items in the response.
      total_count:
        type: integer
        format: int64
        description: The total number of items available.
      items:
        type: array
        items:
          $ref: '#/definitions/Metadata'
paths:
...
  /v1/users/{user_id}:
    get:
      tags:
        - metadata
      operationId: get_user
      summary: Gets A User
      description: Returns a user metadata if it exists.
      parameters:
        - name: user_id
          in: path
          description: The user id to get its metadata.
          required: true
          type: string
          format: id
      responses:
        200:
          description: Returns a user metadata.
          schema:
            $ref: '#/definitions/Metadata'
        404:
          description: User does not exists.
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: Unexpected internal error.
          schema:
            $ref: '#/definitions/ErrorResponse'
    delete:
      tags:
        - metadata
      operationId: delete_user
      summary: Deletes a User
      description: Deletes a user metadata with the given user id.
      parameters:
        - name: user_id
          in: path
          description: The user id to delete its metadata.
          required: true
          type: string
          format: id
      responses:
        200:
          description: Deleted a user metadata.
          schema:
            $ref: '#/definitions/MessageResponse'
        default:
          description: Unexpected internal error.
          schema:
            $ref: '#/definitions/ErrorResponse'
  /v1/users:
    post:
        tags:
          - metadata
        operationId: post_user
        summary: Posts a User
        description: |
          Posts a user metadata.
          Note that this operation completely overrides previous metadata for the id, if it exists.
        parameters:
          - name: user
            in: body
            description: The metadata to be saved. Metadata format has its restrictions.
            required: true
            schema:
              $ref: '#/definitions/Metadata'
        responses:
          201:
            description: Posted a user metadata.
            schema:
              $ref: '#/definitions/MessageResponse'
          400:
            description: Metadata is invalid.
            schema:
              $ref: '#/definitions/DetailedErrorResponse'
          default:
            description: Unexpected internal error.
            schema:
              $ref: '#/definitions/ErrorResponse'
    get:
      tags:
        - metadata
      operationId: get_users
      summary: Gets Users
      description: |
        Get items and total count of items.
        Page and per-page parameters could be set.
      parameters:
        - name: size
          in: query
          description: |
            The number of the users response.
            Defaults to 10. Must be between 1 and 10,000 inclusive.
            This parameter must be string represetation of an integer like "1".
          required: false
          type: integer
          format: int64
        - name: from
          in: query
          description: |
            The number of users to be skipped from the response.
            Defaults to 0. Must be bigger than or equal to 0.
            This parameter must be string represetation of an integer like "1".
          required: false
          type: integer
          format: int64
      responses:
        200:
          description: Get users response.
          schema:
            $ref: '#/definitions/UsersResponse'
        default:
          description: Unexpected internal error.
          schema:
            $ref: '#/definitions/ErrorResponse'
    delete:
      tags:
        - metadata
      operationId: delete_all_users
      summary: Deletes All Users
      description: |
        Warning: Deletes all user metadata from SuggestGrid.
      responses:
        200:
          description: Deleted all user metadata.
          schema:
            $ref: '#/definitions/MessageResponse'
        default:
          description: Unexpected internal error.
          schema:
            $ref: '#/definitions/ErrorResponse'
  /v1/items/{item_id}:
    get:
      tags:
        - metadata
      operationId: get_item
      summary: Gets An Item
      description: Returns an item metadata if it exists.
      parameters:
        - name: item_id
          in: path
          description: The item id to get its metadata.
          required: true
          type: string
          format: id
      responses:
        200:
          description: Returns an item metadata.
          schema:
            $ref: '#/definitions/Metadata'
        404:
          description: Item does not exists.
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: Unexpected internal error.
          schema:
            $ref: '#/definitions/ErrorResponse'
    delete:
      tags:
        - metadata
      operationId: delete_item
      summary: Delete An Item
      description: Deletes an item metadata with the given item id.
      parameters:
        - name: item_id
          in: path
          description: The item id to delete its metadata.
          required: true
          type: string
          format: id
      responses:
        200:
          description: Deleted an item metadata.
          schema:
            $ref: '#/definitions/MessageResponse'
        default:
          description: Unexpected internal error.
          schema:
            $ref: '#/definitions/ErrorResponse'
  /v1/items:
    post:
      tags:
        - metadata
      operationId: post_item
      summary: Posts An Item
      description: |
        Posts an item metadata.
        Note that this operation completely overrides previous metadata for the id, if it exists.
      parameters:
        - name: item
          in: body
          description: The metadata to be saved. Metadata format has its restrictions.
          required: true
          schema:
            $ref: '#/definitions/Metadata'
      responses:
        201:
          description: Posted an item metadata.
          schema:
            $ref: '#/definitions/MessageResponse'
        400:
          description: Metadata is invalid.
          schema:
            $ref: '#/definitions/DetailedErrorResponse'
        default:
          description: Unexpected internal error.
          schema:
            $ref: '#/definitions/ErrorResponse'
    get:
      tags:
        - metadata
      operationId: get_items
      summary: Gets Items
      description: |
        Gets items and total count of items.
        Page and per-page parameters could be set.
      parameters:
        - name: size
          in: query
          description: |
            The number of the users response.
            Defaults to 10. Must be between 1 and 10,000 inclusive.
            This parameter must be string represetation of an integer like "1".
          required: false
          type: integer
          format: int64
        - name: from
          in: query
          description: |
            The number of users to be skipped from the response.
            Defaults to 0. Must be bigger than or equal to 0.
            This parameter must be string represetation of an integer like "1".
          required: false
          type: integer
          format: int64
      responses:
        200:
          description: Get items response.
          schema:
            $ref: '#/definitions/ItemsResponse'
        default:
          description: Unexpected internal error.
          schema:
            $ref: '#/definitions/ErrorResponse'
    delete:
      tags:
        - metadata
      operationId: delete_all_items
      summary: Deletes All Items
      description: |
        Warning: Deletes all item metadata from SuggestGrid.
      responses:
        200:
          description: Deleted all item metadata.
          schema:
            $ref: '#/definitions/MessageResponse'
        default:
          description: Unexpected internal error.
          schema:
            $ref: '#/definitions/ErrorResponse'
  /v1/users/_bulk:
    post:
      tags:
        - metadata
      operationId: post_bulk_users
      summary: Posts Users
      description: |
        Posts user metadata in bulk.
        Note that this operation completely overrides metadata with the same ids, if they exist.
      consumes:
       - text/plain; charset=utf-8
      parameters:
        - name: users
          in: body
          description:  |
            A number of user metadata objects separated with newlines.
            Each user metadata object must have its id field.
            Note that this is not a valid JSON data structure.
            The body size is limited to 10 thousand lines.
          required: true
          schema:
            type: string
      responses:
        201:
          description: Bulk user metadata request executed.
          schema:
            $ref: '#/definitions/BulkPostResponse'
        400:
          description: Body is missing.
          schema:
            $ref: '#/definitions/ErrorResponse'
        413:
          description: Bulk request maximum line count exceeded.
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: Unexpected internal error.
          schema:
            $ref: '#/definitions/ErrorResponse'
  /v1/items/_bulk:
    post:
      tags:
        - metadata
      operationId: post_bulk_items
      summary: Posts Items
      description: |
        Posts item metadata in bulk.
        Note that this operation completely overrides metadata with the same ids, if they exist.
      consumes:
       - text/plain; charset=utf-8
      parameters:
        - name: items
          in: body
          description:   |
            A number of item metadata objects separated with newlines.
            Each item metadata object must have its id field.
            Note that this is not a valid JSON data structure.
            The body size is limited to 10 thousand lines.
          required: true
          schema:
            type: string
      responses:
        201:
          description: Bulk item metadata request executed.
          schema:
            $ref: '#/definitions/BulkPostResponse'
        400:
          description: Body is missing.
          schema:
            $ref: '#/definitions/ErrorResponse'
        413:
          description: Bulk request maximum line count exceeded.
          schema:
            $ref: '#/definitions/ErrorResponse'
        default:
          description: Unexpected internal error.
          schema:
            $ref: '#/definitions/ErrorResponse'
  # Recommendation Paths
...